+++ src/ota.cpp
@@ -180,7 +180,7 @@ int do_ota_update() {
 
     unsigned long timeout = millis();
     while (client.available() == 0) {
-      if ((millis() - timeout) > (RESPONSE_TIMEOUT_MS)) {
+      if ((long)(millis() - timeout) > (RESPONSE_TIMEOUT_MS)) {
         ESP_LOGI(TAG, "Client timeout");
         ota_display(3, " E", "client timeout");
         goto abort;
