+++ src/rcommand.cpp
@@ -298,8 +298,8 @@ void get_config(uint8_t val[]) {
 void get_status(uint8_t val[]) {
   ESP_LOGI(TAG, "Remote command: get device status");
   payload.reset();
-  payload.addStatus(read_voltage(), uptime() / 1000, temperatureRead(),
-                    getFreeRAM(), rtc_get_reset_reason(0),
+  payload.addStatus(read_voltage(), (uint64_t)(uptime() / 1000ULL),
+                    temperatureRead(), getFreeRAM(), rtc_get_reset_reason(0),
                     rtc_get_reset_reason(1));
   SendPayload(STATUSPORT);
 };
